# api/v3/openapi.yaml
openapi: 3.0.0
info:
  title: Kafka HTTP API
  version: 3.0.0
servers:
  - url: "{protocol}://{address}:{port}/v3"
    description: Kafka REST Proxy. See https://docs.confluent.io/current/kafka-rest/index.html.
    variables:
      address:
        default: localhost
        description: Address of the server
      port:
        default: "8082"
        description: Port of the server
      protocol:
        default: http
        description: Protocol for interacting with server
        enum:
          - http
          - https
  - url: "{protocol}://{address}:{port}/kafka/v3"
    description: Kafka REST Proxy as part of Confluent Server.
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
        description: Protocol to communicate with the server
      address:
        default: localhost
        description: Address of the server
      port:
        default: "8090"
        description: Port of the server
paths:
...
  /clusters/{cluster_id}/consumer-groups/{consumer_group_id}/consumers:
    parameters:
      - $ref: '#/components/parameters/ClusterId'
      - $ref: '#/components/parameters/ConsumerGroupId'

    get:
      summary: 'List Consumers'
      description: 'Returns a list of consumers that belong to the specified consumer group.'
      tags:
        - Consumer Group
      responses:
        '200':
          $ref: '#/components/responses/ListConsumersResponse'
  /clusters/{cluster_id}/consumer-groups/{consumer_group_id}/consumers/{consumer_id}:
    parameters:
      - $ref: '#/components/parameters/ClusterId'
      - $ref: '#/components/parameters/ConsumerGroupId'
      - $ref: '#/components/parameters/ConsumerId'

    get:
      summary: 'Get Consumer'
      description: 'Returns the consumer specified by the ``consumer_id``.'
      tags:
        - Consumer Group
      responses:
        '200':
          $ref: '#/components/responses/GetConsumerResponse'
...
components:
  parameters:
...
    ClusterId:
      name: 'cluster_id'
      description: 'The Kafka cluster ID.'
      in: path
      required: true
      schema:
        type: string
      example: 'cluster-1'
...
    ConsumerGroupId:
      name: 'consumer_group_id'
      description: 'The consumer group ID.'
      in: path
      required: true
      schema:
        type: string
      example: 'consumer-group-1'
    ConsumerId:
      name: 'consumer_id'
      description: 'The consumer ID.'
      in: path
      required: true
      schema:
        type: string
      example: 'consumer-1'
...
  schemas:
...
    ConsumerData:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          required:
            - cluster_id
            - consumer_group_id
            - consumer_id
            - client_id
            - assignments
          properties:
            cluster_id:
              type: string
            consumer_group_id:
              type: string
            consumer_id:
              type: string
            instance_id:
              type: string
              nullable: true
            client_id:
              type: string
            assignments:
              $ref: '#/components/schemas/Relationship'
    ConsumerDataList:
      allOf:
        - $ref: '#/components/schemas/ResourceCollection'
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ConsumerData'
...
    Relationship:
      type: object
      required:
        - related
      properties:
        related:
          type: string
...
    Resource:
      type: object
      required:
        - kind
        - metadata
      properties:
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/ResourceMetadata'
    ResourceCollection:
      type: object
      required:
        - kind
        - metadata
      properties:
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/ResourceCollectionMetadata'
    ResourceCollectionMetadata:
      type: object
      required:
        - self
      properties:
        self:
          type: string
        next:
          type: string
          nullable: true
    ResourceMetadata:
      type: object
      required:
        - self
      properties:
        self:
          type: string
        resource_name:
          type: string
          nullable: true
...
  responses:
...
    GetConsumerResponse:
      description: 'The consumer.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConsumerData'
          example:
            kind: 'KafkaConsumer'
            metadata:
              self: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-1'
              resource_name: 'crn:///kafka=cluster-1/consumer-group=consumer-group-1/consumer=consumer-1'
            cluster_id: 'cluster-1'
            consumer_group_id: 'consumer-group-1'
            consumer_id: 'consumer-1'
            instance_id: 'consumer-instance-1'
            client_id: 'client-1'
            assignments:
              related: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-1/assignments'
...
    ListConsumersResponse:
      description: 'The list of consumers.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConsumerDataList'
          example:
            kind: 'KafkaConsumerList'
            metadata:
              self: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers'
              next: null
            data:
              - kind: 'KafkaConsumer'
                metadata:
                  self: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-1'
                  resource_name: 'crn:///kafka=cluster-1/consumer-group=consumer-group-1/consumer=consumer-1'
                cluster_id: 'cluster-1'
                consumer_group_id: 'consumer-group-1'
                consumer_id: 'consumer-1'
                instance_id: 'consumer-instance-1'
                client_id: 'client-1'
                assignments:
                  related: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-1/assignments'
              - kind: 'KafkaConsumer'
                metadata:
                  self: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-2'
                  resource_name: 'crn:///kafka=cluster-1/consumer-group=consumer-group-1/consumer=consumer-2'
                cluster_id: 'cluster-1'
                consumer_group_id: 'consumer-group-1'
                consumer_id: 'consumer-2'
                instance_id: 'consumer-instance-2'
                client_id: 'client-2'
                assignments:
                  related: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-2/assignments'
              - kind: 'KafkaConsumer'
                metadata:
                  self: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-2'
                  resource_name: 'crn:///kafka=cluster-1/consumer-group=consumer-group-1/consumer=consumer-2'
                cluster_id: 'cluster-1'
                consumer_group_id: 'consumer-group-1'
                consumer_id: 'consumer-2'
                instance_id: 'consumer-instance-2'
                client_id: 'client-2'
                assignments:
                  related: 'http://localhost:9391/v3/clusters/cluster-1/consumer-groups/consumer-group-1/consumers/consumer-2/assignments'
...